Fix cyclic dependencies
