+++ application/ui/src/routes/layout.tsx
@@ -1,7 +1,7 @@
 import { Flex, Grid, Tabs, View } from '@geti/ui';
 import { Outlet, useLocation } from 'react-router';
 
-import { ProjectsListPanel } from './features/inspect/projects-management/projects-list-panel.component';
+import { ProjectsListPanel } from './../features/inspect/projects-management/projects-list-panel.component';
 
 const Header = () => {
     return (
